added comments and backward compatibility support
